feat(papergen.py): Initial working .md parser to generate the final paper
Does not do much but concatenate the .md files from the wiki as specified by
the main page "Final-Report.md" and print the result to the command line.
The parsing itself is straight forward.

```
${include:{Some-cool-file.md}}

${include:{Another-file.md}}
```

will be transformed to

```
Some cool file
==============
This is the contents of the cool file.

Another file
============
Moar text!
```

Closes #40